Parameterize test262 config.
This will allow it to be re-used over in the runner repo for basic CI
testing.
